finished group sms
Former-commit-id: f45daa18758c6a0a13b695332c6a33ddaa231ccc

+++ lib/src/messages/messages_list.dart
@@ -390,7 +390,10 @@ class _SMSConversationSummaryViewState
             ),
           ),
           onDismissed: (DismissDirection direction) {
-            _fusionConnection.conversations.deleteConversation(_convo, _departmentId);
+            if(_deleteConvo != null){
+              _fusionConnection.conversations.deleteConversation(_convo, _departmentId);
+              _deleteConvo(_convo,null);
+            }
           },
           confirmDismiss: (DismissDirection direction) async {
             return await showDialog(
